+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 0.2.2
 -----
 
-- Add dodecahedron project logo to Sphinx documentation sidebar.
+- Add dodecahedron project logo to README and repository.
 
 0.2.1
 -----